requirements.replicate.txt
networkx
torch==2.5.1
torchvision==0.20.1
einops
numpy==1.26.4
opencv-python
Pillow
scikit-image
pyclipper
shapely
requests
tqdm
aiohttp
freetype-py
kornia
huggingface_hub
transformers
sentencepiece
ctranslate2
safetensors
timm
omegaconf
python-dotenv
colorama
pydantic==2.5.0
protobuf<6.0.0,>=3.20.2
regex
editdistance
py3langid==0.2.2
langcodes[data]
langdetect
python-bidi
arabic-reshaper
pyhyphen
marshmallow
aiofiles
aioshutil
nest-asyncio
cython
pydensecrf@https://github.com/lucasb-eyer/pydensecrf/archive/refs/heads/master.zip
